moved tests to new folder
