+++ backend/pytest.ini
@@ -5,3 +5,4 @@ env =
 	SECRET_KEY=secret
 	OBJECT_URL=https://smart-speech-media.s3.amazonaws.com
 	AWS_REGION=us-east-1
+testpaths=./src/routers/tests ./src
