+++ backend/src/routers/tests/test_s3.py
@@ -7,10 +7,10 @@ from moto import mock_s3
 
 from unittest.mock import patch
 
-from .s3 import router
+from ..s3 import router
 
-from .aws_constants import UPLOAD_TO_S3_ROUTE, SAMPLE_BASE64_IMAGE, SAMPLE_UUIDv4
-from .aws_constants import OBJECT_URL_ENV_VAR, AWS_REGION_ENV_VAR, BUCKET_NAME_ENV_VAR
+from ..aws_constants import UPLOAD_TO_S3_ROUTE, SAMPLE_BASE64_IMAGE, SAMPLE_UUIDv4
+from ..aws_constants import OBJECT_URL_ENV_VAR, AWS_REGION_ENV_VAR, BUCKET_NAME_ENV_VAR
 
 client = TestClient(app=router)
